pyproject.toml
[tool.poetry]
name = "redeval"
version = "0.1.0"
description = "Auditing and Label-free Evaluation for LLMs"
authors = ["Christos Ziakas"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.28.0"
pandas = "^2.1.0"
python-dotenv = "^1.0.0"
datasets = "^2.14.5"
scikit-learn = "^1.3.0"
llama-index = "^0.8.40"
pre-commit = "^3.4.0"
transformers = "^4.34.0"
torch = "^2.1.0"
torchvision = "^0.16.0"
replicate = "^0.15.4"



[tool.poetry.group.dev.dependencies]
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
